Update monster.h
변수 double로 변경 시켰습니다.

+++ main project/monster.h
@@ -4,12 +4,12 @@
 using namespace std;
 
 typedef struct mob {
-    float hp, mp, atk, def, speed;//선공 선별(speed)
+    double hp, mp, atk, def, speed;//선공 선별(speed)
     string ettype; //몹종류 -> normal,elite,boss
     string kind; //종족
 
-    mob(string kind, float hp = 0, float mp = 0, float atk = 0,
-        float def = 0, float speed = 1, string ettype = "normal") //default 매개변수 맨끝
+    mob(string kind, double hp = 0, double mp = 0, double atk = 0,
+        double def = 0, double speed = 1, string ettype = "normal") //default 매개변수 맨끝
     {
         //ex) mob Slime_Normal = { "slime",100,20,10,5,"normal" };
 
@@ -27,19 +27,19 @@ typedef struct mob {
         cout << "종족 : " << kind << " [" << ettype << "] " << endl;
         cout << "HP : " << hp << endl;
     }
-    void hp_plus(float a)
+    void hp_plus(double a)
     {
         this->hp += a;
     }
-    void mp_plus(float a)
+    void mp_plus(double a)
     {
         this->mp += a;
     }
-    void atk_plus(float a)
+    void atk_plus(double a)
     {
         this->atk += a;
     }
-    void def_plus(float a)
+    void def_plus(double a)
     {
         this->def += a;
     }
